add security on resource records

+++ src/main/java/com/softserve/rms/service/implementation/ResourceRecordServiceImpl.java
@@ -7,6 +7,7 @@ import com.softserve.rms.entities.ResourceRecord;
 import com.softserve.rms.entities.ResourceTemplate;
 import com.softserve.rms.entities.User;
 import com.softserve.rms.exceptions.NotFoundException;
+import com.softserve.rms.exceptions.PermissionException;
 import com.softserve.rms.exceptions.resourseTemplate.ResourceTemplateIsNotPublishedException;
 import com.softserve.rms.repository.ResourceRecordRepository;
 import com.softserve.rms.service.ResourceRecordService;
@@ -133,6 +134,7 @@ public class ResourceRecordServiceImpl implements ResourceRecordService {
             throws NotFoundException {
         checkIfResourceTemplateIsPublished(tableName);
         ResourceRecord resourceRecord = findById(tableName, id);
+        checkPermissionOnResource(resourceRecord);
         if (resourceRecordSaveDTO.getName() != null) {
             resourceRecord.setName(resourceRecordSaveDTO.getName());
         }
@@ -154,6 +156,7 @@ public class ResourceRecordServiceImpl implements ResourceRecordService {
     public void delete(String tableName, Long id) throws NotFoundException {
         checkIfResourceTemplateIsPublished(tableName);
         ResourceRecord resourceRecord = findById(tableName, id);
+        checkPermissionOnResource(resourceRecord);
         if (resourceRecord.getPhotosNames() != null) {
             deletePhotosFromS3(resourceRecord.getPhotosNames());
         }
@@ -177,6 +180,7 @@ public class ResourceRecordServiceImpl implements ResourceRecordService {
     public void changePhoto(MultipartFile files,
                             String tableName, Long id) {
         ResourceRecord resourceRecord = findById(tableName, id);
+        checkPermissionOnResource(resourceRecord);
         String photoName = fileStorageService.uploadFile(files);
         if (resourceRecord.getPhotosNames() != null) {
             resourceRecord.setPhotosNames(resourceRecord.getPhotosNames() + photoName + ',');
@@ -194,6 +198,7 @@ public class ResourceRecordServiceImpl implements ResourceRecordService {
     @Override
     public void deleteAllPhotos(String tableName, Long id) {
         ResourceRecord resourceRecord = findById(tableName, id);
+        checkPermissionOnResource(resourceRecord);
         deletePhotosFromS3(resourceRecord.getPhotosNames());
         resourceRecord.setPhotosNames(null);
         resourceRecordRepository.update(tableName, id, resourceRecord);
@@ -207,6 +212,7 @@ public class ResourceRecordServiceImpl implements ResourceRecordService {
     @Override
     public void deletePhoto(String tableName, Long id, String photo) {
         ResourceRecord resourceRecord = findById(tableName, id);
+        checkPermissionOnResource(resourceRecord);
         Stream.of(resourceRecord.getPhotosNames().split(",")).
                 filter(p -> p.equals(photo)).
                 forEach(q -> fileStorageService.deleteFile(q));
@@ -238,4 +244,10 @@ public class ResourceRecordServiceImpl implements ResourceRecordService {
         return (result.toString());
     }
 
+    private void checkPermissionOnResource(ResourceRecord resourceRecord ) {
+        Principal principal = SecurityContextHolder.getContext().getAuthentication();
+        if(!resourceRecord.getUser().getEmail().equals(principal.getName())) {
+            throw new PermissionException(ErrorMessage.ACCESS_DENIED.getMessage());
+        }
+    }
 }
